Use content provider to get source.
